// Frontend/src/app/modules/ct-client/model/takealot-content-response.ts
/**
 * ChangeTrackerAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TakealotContentResponse { 
    id?: string;
    name?: string | null;
    lastUpdated?: Date;
    url?: string | null;
    productIdentifier?: string | null;
    inStock?: boolean;
}